requirements.txt
fastapi
uvicorn[standard]
langgraph
langchain-groq
psycopg2-binary
passlib[bcrypt]
bcrypt==4.0.1
python-jose[cryptography]
PyJWT
pydantic
python-multipart
jinja2
pandas
supabase